Conflicts:
	AccelerometerPlay/SimulationView.cs
	ApiDemo/Animation/Transition3d.cs
	ApiDemo/ApiDemo.cs
	ApiDemo/Tutorials/GalleryTutorial.cs
	ApiDemo/Tutorials/GridViewTutorial.cs
	ApiDemo/Tutorials/ListViewTutorial.cs
	ApiDemo/Tutorials/SpinnerTutorial.cs
	BluetoothChat/BluetoothChat.cs
	BluetoothChat/DeviceListActivity.cs
	ContentControls/ContentControls/ContentControls.csproj
	ContentControls/ContentControls/Home_Adapter.cs
	ContentControls/ContentControls/Screens/Search/AutoCompleteTextViewScreen.cs
	GLCube/PaintingView.cs
	GLDiagnostics/GLDiag.csproj
	GLDiagnostics/GLDiag.sln
	GLDiagnostics/PaintingView.cs
	GLDiagnostics/Resources/layout/main.xml
	GLTriangle20/PaintingView.cs
	JetBoy/Classes/JetBoyView.cs
	LiveWallpaperDemo/CubeWallpaper.cs
	MapsDemo/RotateView.cs
	NotePad/Resources/Resource.Designer.cs
	SanityTests/Hello.cs
	Snake/Resources/values/strings.xml
	StandardControls/StandardControls/Screens/ContentControls/GalleryScreen.cs
	StandardControls/StandardControls/Screens/ContentControls/GridViewScreen.cs
	TablesAndCellStyles/TablesAndCellStyles/Screens/AccessoryStyles/SimpleListItemChecked.cs
	TablesAndCellStyles/TablesAndCellStyles/Screens/AccessoryStyles/SimpleListItemMultipleChoice.cs
	TablesAndCellStyles/TablesAndCellStyles/Screens/AccessoryStyles/SimpleListItemSingleChoice.cs
	TablesAndCellStyles/TablesAndCellStyles/Screens/CustomTables/GradientBackground/GradientBackground.cs
	TablesAndCellStyles/TablesAndCellStyles/Screens/CustomTables/LabelledSectionsIndexer/LabelledSectionsIndexer.cs
	TexturedCube/PaintingView.cs
	build.sh
	pre-commit.posix
	pre-commit.windows
	validate.posix

+++ ApiDemo/App/ActionBarDisplayOptions.cs
@@ -90,7 +90,7 @@ namespace MonoDroid.ApiDemo.App
 			var bar = ActionBar;
 
 			var change = bar.DisplayOptions ^ flag;
-			ActionBar.SetDisplayOptions (change, (int)flag);
+			ActionBar.SetDisplayOptions (change, flag);
 		}
 
 		private void CycleGravity (object sender, EventArgs e)
